Default to creating a WebViewController if no custom view is found
If the PageRegistry can't find a custom UIViewController implementation
when looking for a page by name, it will create a WebViewController to
display the shell page.

This change makes it so the iOS app will implicitly show the shell
views without any work from the developer. To use a native view, you
just need to create a UIViewController and the WebViewController won't
be used anymore.