Don't capitalize the page names that are stored
The handling of page name casing is somewhat inconsistent, and this
causes issues when we pass the name directly to the WebViewController.

The registry was capitalizing the page name, and the WebViewController
pageName method is used to call the window.pageViews.pageName.render
function; so when you pass the capitalized page name directly into the
WebViewController it ends up trying to call an undefined object on
pageViews.

We should probably normalise all names to a standard format, but until
then this commit changes it so we don't touch the casing apart from
when we try to find the ViewController by name.