Be safer about executing javascript in the web views
Whenever javascript is executed in the web views, namely to render
views or fire events, there's the possibility the call might throw an
exception; when an exception is uncaught, the whole javascript
execution fails and the app becomes useless.

This change wraps every call to the javascript runtime in a try...catch
block and returns any unhandled exceptions to the caller. If the result
contains an exception, it's rethrown as an NSException.